+++ programs/fixed-term/src/lib.rs
@@ -171,7 +171,7 @@ pub mod jet_fixed_term {
 
     /// Modify a `Market` account
     /// Authority use only
-    pub fn modify_market(ctx: Context<ModifyMarket>, data: Vec<u8>, offset: usize) -> Result<()> {
+    pub fn modify_market(ctx: Context<ModifyMarket>, data: Vec<u8>, offset: u32) -> Result<()> {
         instructions::modify_market::handler(ctx, data, offset)
     }
